Don't check for failback flush when `$wpdb->options` isn't yet available
On multisite, the object cache is initialized before `$wpdb` is fully
configured. `$wpdb` is only fully configured after WordPress figures out
which blog to load. When it does so, it reinitializes the object cache.

Because the object cache is reinitialized, we can skip the failback
flush check when `$wpdb->options` isn't yet available.